tests/requirements.txt
# Devops tools
flake8==3.9.2
flake8-html==0.4.1
pylint==2.8.3
pylint-json2html==0.3.0
bandit==1.7.0

# Unit tests tools
pytest==6.2.4
pytest-django==4.4.0
pytest-mock==3.6.1
faker==8.8.2